Move platform files to proper folders
