+++ src/NzbDrone.Mono.Test/DiskProviderTests/DiskProviderFixture.cs
@@ -2,6 +2,7 @@
 using Mono.Unix;
 using NUnit.Framework;
 using NzbDrone.Common.Test.DiskTests;
+using NzbDrone.Mono.Disk;
 
 namespace NzbDrone.Mono.Test.DiskProviderTests
 {
