+++ src/NzbDrone.Mono.Test/DiskProviderTests/FreeSpaceFixture.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using NzbDrone.Common.Test.DiskTests;
+using NzbDrone.Mono.Disk;
 
 namespace NzbDrone.Mono.Test.DiskProviderTests
 {
