+++ src/NzbDrone.Windows.Test/DiskProviderTests/FreeSpaceFixture.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
 using NzbDrone.Common.Test.DiskTests;
+using NzbDrone.Windows.Disk;
 
 namespace NzbDrone.Windows.Test.DiskProviderTests
 {
